requirements.txt
python-telegram-bot
python-dotenv
scikit-learn
